+++ src/main/java/powerup/v1/repositories/MissaoConfigRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 import powerup.v1.entities.MissaoConfig;
 
 @Repository
-public interface MissaoConfigRepository extends JpaRepository<MissaoConfig, Long> {
+public interface MissaoConfigRepository extends JpaRepository<MissaoConfig, Integer> {
 }
